+++ Microsoft.Azure.Cosmos/src/CosmosElements/Numbers/CosmosFloat64.EagerCosmosFloat64.cs
@@ -1,6 +1,9 @@
-﻿//------------------------------------------------------------
+//------------------------------------------------------------
 // Copyright (c) Microsoft Corporation.  All rights reserved.
 //------------------------------------------------------------
+
+// This is auto-generated code. Modify: CosmosNumberCodeGenerator.tt: 107
+
 namespace Microsoft.Azure.Cosmos.CosmosElements.Numbers
 {
 #if INTERNAL
@@ -27,8 +30,4 @@ namespace Microsoft.Azure.Cosmos.CosmosElements.Numbers
             }
         }
     }
-#if INTERNAL
-#pragma warning restore SA1601 // Partial elements should be documented
-#pragma warning restore CS1591 // Missing XML comment for publicly visible type or member
-#endif
 }