+++ Microsoft.Azure.Cosmos/src/Resource/Container/ContainerCore.Items.cs
@@ -366,6 +366,7 @@ namespace Microsoft.Azure.Cosmos
             {
                 queryIterator = QueryIterator.Create(
                     client: this.queryClient,
+                    clientContext: this.ClientContext,
                     sqlQuerySpec: queryDefinition.ToSqlQuerySpec(),
                     continuationToken: continuationToken,
                     feedRangeInternal: feedRangeInternal,
@@ -607,6 +608,7 @@ namespace Microsoft.Azure.Cosmos
 
             return QueryIterator.Create(
                 client: this.queryClient,
+                clientContext: this.ClientContext,
                 sqlQuerySpec: sqlQuerySpec,
                 continuationToken: continuationToken,
                 feedRangeInternal: feedRange,