+++ Microsoft.Azure.Cosmos/src/Resource/Settings/ThroughputProperties.cs
@@ -70,14 +70,14 @@ namespace Microsoft.Azure.Cosmos
         public int? Throughput
         {
             get => this.Content?.OfferThroughput;
-            private set => this.Content = OfferContentProperties.CreateFixedOfferConent(value.Value);
+            private set => this.Content = OfferContentProperties.CreateManualOfferConent(value.Value);
         }
 
         /// <summary>
         /// The maximum throughput the autoscale will scale to.
         /// </summary>
         [JsonIgnore]
-#if INTERNAL
+#if PREVIEW
         public
 #else
         internal
@@ -88,40 +88,42 @@ namespace Microsoft.Azure.Cosmos
         /// The amount to increment if the maximum RUs is getting throttled.
         /// </summary>
         [JsonIgnore]
-#if INTERNAL
-        public
-#else
-        internal
-#endif
-        int? AutoUpgradeMaxThroughputIncrementPercentage => this.Content?.OfferAutoscaleSettings?.AutoscaleAutoUpgradeProperties?.ThroughputProperties?.IncrementPercent;
+        internal int? AutoUpgradeMaxThroughputIncrementPercentage => this.Content?.OfferAutoscaleSettings?.AutoscaleAutoUpgradeProperties?.ThroughputProperties?.IncrementPercent;
 
         /// <summary>
-        /// The Throughput properties for autoscale provisioned throughput offering
+        /// The Throughput properties for manual provisioned throughput offering
         /// </summary>
         /// <param name="throughput">The current provisioned throughput for the resource.</param>
-        /// <returns>Returns a ThroughputProperties for fixed throughput</returns>
-#if INTERNAL
+        /// <returns>Returns a ThroughputProperties for manual throughput</returns>
+#if PREVIEW
         public
 #else
         internal
 #endif
-        static ThroughputProperties CreateFixedThroughput(int throughput)
+        static ThroughputProperties CreateManualThroughput(int throughput)
         {
-            return new ThroughputProperties(OfferContentProperties.CreateFixedOfferConent(throughput));
+            return new ThroughputProperties(OfferContentProperties.CreateManualOfferConent(throughput));
         }
 
         /// <summary>
         /// The Throughput properties for autoscale provisioned throughput offering
         /// </summary>
-        /// <param name="maxAutoscaleThroughput">The staring maximum throughput the resource can scale to.</param>
-        /// <param name="autoUpgradeMaxThroughputIncrementPercentage">The percentage to increase the maximum value if the maximum is being throttled.</param>
+        /// <param name="maxAutoscaleThroughput">The maximum throughput the resource can scale to.</param>
         /// <returns>Returns a ThroughputProperties for autoscale provisioned throughput</returns>
-#if INTERNAL
+#if PREVIEW
         public
 #else
         internal
 #endif
-        static ThroughputProperties CreateAutoscaleProvionedThroughput(
+        static ThroughputProperties CreateAutoscaleThroughput(
+            int maxAutoscaleThroughput)
+        {
+            return new ThroughputProperties(OfferContentProperties.CreateAutoscaleOfferConent(
+                startingMaxThroughput: maxAutoscaleThroughput,
+                autoUpgradeMaxThroughputIncrementPercentage: null));
+        }
+
+        internal static ThroughputProperties CreateAutoscaleThroughput(
             int maxAutoscaleThroughput,
             int? autoUpgradeMaxThroughputIncrementPercentage = null)
         {